pyproject.toml
[project]
name = "blackpdf"
version = "0.1.0"
description = "Simple PDF color inverter - white background to black, white text"
requires-python = ">=3.8"
dependencies = [
    "Flask>=3.0.0",
    "PyMuPDF>=1.23.8", 
    "Pillow>=10.1.0",
    "python-dotenv>=1.0.0"
]

[tool.hatch.build.targets.wheel]
include = [
    "*.py",
    "templates/*",
    ".env"
]